+++ subprojects/lua-cjson/scripts/lua2json.sh
@@ -1,0 +1,2 @@
+BASEDIR=$(dirname $0)
+exec "$BASEDIR/@BIN@" "$BASEDIR/../lib/luarocks/rocks-@LUA_VER@/lua-cjson/@ROCKSPEC_VER@/bin/lua2json.lua" "$@"